fix ../../include/fast_io_hosted/file_loaders/posix_file_loader.h:434:61: warning: implicit conversion changes signedness: 'long' to 'std::size_t' (aka 'unsigned long') [-Wsign-conversion]                 posix_unload_address<allocation>(address_begin,address_end-address_begin);                 ~~~~~~~~~~~~~~~~~~~~                           ~~~~~~~~~~~^~~~~~~~~~~~~~ cap.cc:13:30: note: in instantiation of member function 'fast_io::details::posix_file_loader_impl<false>::~posix_file_loader_impl' requested here         fast_io::native_file_loader nf(argv[1]);                                     ^ -Wall -Wextra -Wpedantic -Wmisleading-indentation -Wunused -Wuninitialized -Wshadow -Wconversion
for clang

+++ include/fast_io_hosted/file_loaders/posix_file_loader.h
@@ -431,7 +431,7 @@ public:
 	}
 	~posix_file_loader_impl()
 	{
-		posix_unload_address<allocation>(address_begin,address_end-address_begin);
+		posix_unload_address<allocation>(address_begin,static_cast<std::size_t>(address_end-address_begin));
 	}
 };
